Fix gameplay

+++ src/states/PlayGameState.h
@@ -16,7 +16,7 @@ class PlayGameState : public BaseState { //GameState<GameContext, GameStateType>
 
     uint8_t counter = 0;
     uint8_t puffIndex = 0;
-    uint8_t victimCountdown = 200;
+    uint16_t victimCountdown = 200;
     uint8_t victimDelay = 1;
     uint8_t victimLevel = 0;
